Cleaned up package structure by moving command related packages from "com.gmail.inverseconduit" into "com.gmail.inverseconduit.commands.providers"
