+++ src/main/java/com/gmail/inverseconduit/commands/providers/javadoc/JavadocDao.java
@@ -1,4 +1,4 @@
-package com.gmail.inverseconduit.javadoc;
+package com.gmail.inverseconduit.commands.providers.javadoc;
 
 import java.io.IOException;
 import java.nio.file.DirectoryStream;
