+++ src/main/java/com/gmail/inverseconduit/commands/providers/javadoc/LibraryZipFile.java
@@ -1,4 +1,4 @@
-package com.gmail.inverseconduit.javadoc;
+package com.gmail.inverseconduit.commands.providers.javadoc;
 
 import java.io.IOException;
 import java.io.InputStream;
